Move tabs to separate folder
