+++ src/cvasl_gui/tabs/compare.py
