+++ src/cvasl_gui/tabs/data_inspection.py
